changed double to BidDecimal in api

+++ open-meteo-sdk/src/main/java/com/kolosov/openmeteosdk/api/OpenMeteoAPIClient.java
@@ -3,6 +3,7 @@ package com.kolosov.openmeteosdk.api;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.service.annotation.GetExchange;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 //TODO Deal with constant RequestParams
@@ -12,15 +13,23 @@ public interface OpenMeteoAPIClient {
 
     @GetExchange(url = "/forecast")
     OpenMeteoResponse getRawForecast(
-            @RequestParam("latitude") double latitude,
-            @RequestParam("longitude") double longitude,
+            @RequestParam("latitude") BigDecimal latitude,
+            @RequestParam("longitude") BigDecimal longitude,
             @RequestParam("hourly") List<String> hourly,
             @RequestParam(value = "timezone", defaultValue = "Asia/Bangkok") String timezone
     );
 
-    default OpenMeteoResponse getRawForecast(double latitude,double longitude) {
-        List<String> hourly = List.of("precipitation", "temperature_2m", "apparent_temperature", "relative_humidity_2m",
-                "cloud_cover", "precipitation_probability", "wind_speed_10m", "wind_gusts_10m");
+    default OpenMeteoResponse getRawForecast(BigDecimal latitude, BigDecimal longitude) {
+        List<String> hourly = List.of(
+                "precipitation",
+                "temperature_2m",
+                "apparent_temperature",
+                "relative_humidity_2m",
+                "cloud_cover",
+                "precipitation_probability",
+                "wind_speed_10m",
+                "wind_gusts_10m"
+        );
         return getRawForecast(latitude, longitude, hourly, "Asia/Bangkok");
     }
 }
